backend/requirements.txt
fastapi
uvicorn
sqlalchemy
python-multipart
Pillow
reportlab
